requirements.txt
# Core Django
Django
djangorestframework

# CORS handling
django-cors-headers

# Environment variables
python-dotenv

# AI Integration - Google Gemini
google-generativeai

# Document processing
PyMuPDF
pypdf
python-pptx

# Text-to-Speech
gTTS

# Database (PostgreSQL for production)
psycopg2-binary

# Production server
gunicorn

# Utilities (just in case)
Pillow
requests